+++ crates/goose/src/agents/subagent_execution_tool/task_types.rs
@@ -164,78 +164,3 @@ pub struct ExecutionResponse {
     pub results: Vec<TaskResult>,
     pub stats: ExecutionStats,
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn test_task_type_serialization() {
-        // Test that TaskType serializes to the expected string format
-        assert_eq!(
-            serde_json::to_string(&TaskType::TextInstruction).unwrap(),
-            "\"text_instruction\""
-        );
-        assert_eq!(
-            serde_json::to_string(&TaskType::InlineRecipe).unwrap(),
-            "\"inline_recipe\""
-        );
-        assert_eq!(
-            serde_json::to_string(&TaskType::SubRecipe).unwrap(),
-            "\"sub_recipe\""
-        );
-    }
-
-    #[test]
-    fn test_task_type_deserialization() {
-        // Test that strings deserialize to the correct TaskType variants
-        assert_eq!(
-            serde_json::from_str::<TaskType>("\"text_instruction\"").unwrap(),
-            TaskType::TextInstruction
-        );
-        assert_eq!(
-            serde_json::from_str::<TaskType>("\"inline_recipe\"").unwrap(),
-            TaskType::InlineRecipe
-        );
-        assert_eq!(
-            serde_json::from_str::<TaskType>("\"sub_recipe\"").unwrap(),
-            TaskType::SubRecipe
-        );
-    }
-
-    #[test]
-    fn test_task_serialization_with_enum() {
-        let task = Task {
-            id: "test-id".to_string(),
-            task_type: TaskType::InlineRecipe,
-            payload: json!({"recipe": "test"}),
-        };
-
-        let serialized = serde_json::to_value(&task).unwrap();
-        assert_eq!(serialized["id"], "test-id");
-        assert_eq!(serialized["task_type"], "inline_recipe");
-        assert_eq!(serialized["payload"]["recipe"], "test");
-    }
-
-    #[test]
-    fn test_task_deserialization_with_string() {
-        // Test backward compatibility - JSON with string task_type should deserialize
-        let json_str = r#"{
-            "id": "test-id",
-            "task_type": "sub_recipe",
-            "payload": {"sub_recipe": {"name": "test"}}
-        }"#;
-
-        let task: Task = serde_json::from_str(json_str).unwrap();
-        assert_eq!(task.id, "test-id");
-        assert_eq!(task.task_type, TaskType::SubRecipe);
-    }
-
-    #[test]
-    fn test_task_type_display() {
-        assert_eq!(TaskType::TextInstruction.to_string(), "text_instruction");
-        assert_eq!(TaskType::InlineRecipe.to_string(), "inline_recipe");
-        assert_eq!(TaskType::SubRecipe.to_string(), "sub_recipe");
-    }
-}
